# openspec/api-spec.yaml
openapi: 3.0.3
info:
  title: FamilyFinances API
  version: v1
  description: |
    As-built OpenAPI specification for FamilyFinances.

    Scope covered:
    - REST API endpoints hosted by `FamilyFinances.Api` (`/api/v1/...`, `/health`)
    - Session endpoints hosted by `FamilyFinances.Web` (`/auth/session`)

    Notes:
    - Authentication in API endpoints is JWT Bearer-based.
    - Web session endpoints proxy login and manage the `ff_access_token` HttpOnly cookie.
  contact:
    name: FamilyFinances Maintainers
    url: https://github.com/carlos-aisa/FamilyFinances
servers:
  - url: http://localhost:5084
    description: Development API server (`FamilyFinances.Api`) and ZIP production default API endpoint
  - url: https://localhost:7249
    description: Development HTTPS API profile (`FamilyFinances.Api`)
...
paths:
...
  /api/v1/accounts/{id}/rename:
    patch:
      tags: [Accounts]
      summary: Rename account
      operationId: Accounts_Rename
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameAccountRequest'
      responses:
        '204':
          description: Account renamed.
        '404':
          description: Account not found (response body may be empty).
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/accounts/{id}:
    delete:
      tags: [Accounts]
      summary: Delete account
      operationId: Accounts_Delete
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '204':
          description: Account deleted.
        '404':
          description: Account not found (response body may be empty).
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/account-groups/{id}:
    get:
      tags: [AccountGroups]
      summary: Get account group by id
      operationId: AccountGroups_GetById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      responses:
        '200':
          description: Account group details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountGroupDetailsDto'
        '404':
          description: Account group not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [AccountGroups]
      summary: Partially update account group dashboard settings
      operationId: AccountGroups_Update
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAccountGroupDashboardPinnedRequest'
      responses:
        '204':
          description: Account group updated.
        '404':
          description: Account group not found.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [AccountGroups]
      summary: Delete account group
      operationId: AccountGroups_Delete
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      responses:
        '204':
          description: Account group deleted.
        '404':
          description: Account group not found (response body may be empty).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/account-groups/{id}/accounts/{accountId}:
    post:
      tags: [AccountGroups]
      summary: Add account to group
      operationId: AccountGroups_AddAccount
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
        - $ref: '#/components/parameters/AccountIdPath2'
      responses:
        '204':
          description: Membership added (idempotent if already present).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [AccountGroups]
      summary: Remove account from group
      operationId: AccountGroups_RemoveAccount
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
        - $ref: '#/components/parameters/AccountIdPath2'
      responses:
        '204':
          description: Membership removed (idempotent).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/account-groups/{id}/rename:
    patch:
      tags: [AccountGroups]
      summary: Rename account group
      operationId: AccountGroups_Rename
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameAccountGroupRequest'
      responses:
        '204':
          description: Account group renamed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Account group not found (response body may be empty).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieToken:
      type: apiKey
      in: cookie
      name: ff_access_token
      description: Token cookie used by the web host session flow.
  parameters:
    AccountIdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    AccountIdPath2:
      in: path
      name: accountId
      required: true
      schema:
        type: string
        format: uuid
...
    GroupIdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Unauthorized.
    Forbidden:
      description: Forbidden.
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server error (response body contract may be empty or framework-defined).
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ProblemDetails:
      type: object
      description: RFC 7807 problem details payload.
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: true
...
    AccountNature:
      type: integer
      format: int32
      enum: [1, 2, 3, 4, 5]
      description: 'Numeric enum: 1=Asset, 2=Liability, 3=Income, 4=Expense, 5=Equity.'
    AccountKind:
      type: integer
      format: int32
      enum: [1, 2, 3, 4, 5, 6, 7, 8, 9, 99]
      description: 'Numeric enum: 1=Checking, 2=Savings, 3=CreditCard, 4=Cash, 5=Investment, 6=ExpenseCategory, 7=IncomeSource, 8=Mortgage, 9=Loan, 99=Other.'
...
    RenameAccountRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
...
    RenameAccountGroupRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
    SetAccountGroupDashboardPinnedRequest:
      type: object
      properties:
        isDashboardPinned:
          type: boolean
      required: [isDashboardPinned]
...
    AccountRefDto:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        name:
          type: string
        nature:
          $ref: '#/components/schemas/AccountNature'
        kind:
          $ref: '#/components/schemas/AccountKind'
      required: [accountId, name, nature, kind]
    AccountGroupDetailsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountRefDto'
        isDashboardPinned:
          type: boolean
      required: [id, name, description, accounts, isDashboardPinned]
...